moved addWorkspace to WorkspaceStripper
